+++ spice/run_all.sh
@@ -22,5 +22,6 @@ run ac_mrb.cir
 run tran_tremolo_lfo.cir
 run ac_power_amp_lm1875.cir
 run tran_reverb_mixer.cir
+run ac_tonestack.cir
 
 exit "$fail"